Add files via upload
These shell scripts help in preparing input files for DFT calculations by the Gaussian package. Step 1 uses a general description of a metal node structure (from a series of metal-organic frameworks) and replaces its generic "Me" metal atoms with the metal atom of choice. To consider all potential spin states, step 2, uses the data about the number of unpaired electrons (see the *.txt file) and generates the input files for optimization and frequency calculations on the metal node structures at various spin states. The most important results can be extracted in Step 3 to see what spin state is stable, lowest in energy and without spin contamination. That spin state, would be the ground spin state for that metal node. In Step 4, the Gaussian input files can be prepared for a large number of organic ligand structures.